EC-EFS: Remove version check for EFS_VERIFY command
EFS_VERIFY command is given a two-byte ID (0x11E). So,
version 0 of GET_CMD_VERSIONS can't answer the query. This results
in successful verification always.

This patch removes the check for EFS_VERIFY version. If the EC does
not support the command, it'll return INVALID_COMMAND. This is
treated as sucess just as before.

BUG=b:71719323
BRANCH=none
TEST=Verification failure results in recovery screen.

Change-Id: I2e1f2623cd816d54b68f3528d3fdc5c8f3151212
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/930002