grunt: Tell DA7219 to configure its BCLK/WCLK as inputs
On cold reset, the DA7219's DAI CLK is disabled, in other words, its BCLK
and WCLK pins are configured as inputs.
Once the system boots, however, the kernel will reconfigure DA7219
to drive BCLK and WCLK as outputs.
Therefore, on a subsequent warm boot, deptcharge must request that the
DA7219 reconfigure its BCLK and WCLK as inputs again to avoid fighting
with it when trying to drive out bit-bang i2s to support the devbeep.

BUG=b:69407112
TEST=Cold boot grunt in dev mode with GBB flag DEV_SCREEN_SHORT_DELAY disabled
 => hear dev beep
 => Once booted, can play audio appropriate audio patches applied
TEST=Warm reboot grunt
 => hear dev beep again
 => Once rebooted, can still play audio appropriate audio patches applied
BRANCH=None

Change-Id: I9781588873dcbb7df1a22dee67e09e169134bef0
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/897937