vboot/screens: Remove logo liftup spacing
The Chrome logo looks a little too high above the divider line for my
tastes right now, especially compared to the language descriptor to the
right of it. The logo image already has some white padding on the bottom
in the image itself, which looks like more than enough on its own.
Remove the extra padding.

BRANCH=None
BUG=b:65479973
TEST=Booted Scarlet, was more aesthetically pleased.

Change-Id: If376ee35088c19dd86fbc6b8673d1cbce1fbf940
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/903241
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>