vboot/screens: Add new help text for detachable menus
This patch adds the three new help messages we carefully bikeshedded to
the DEV_WARNING, TO_NORM and TO_DEV screens, as well as the common "how
to use the menu" help text on the bottom of each menu. It also makes the
pagination of the language menu prettier.

CQ-DEPEND=CL:903237
BRANCH=None
BUG=None
TEST=Booted on Scarlet

Change-Id: I07e67a76b209160878c7ed80d04068884fdc14dc
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/903364
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>