image: Remove FMAP_OFFSET Kconfig, read it from sysinfo
We've been passing the FMAP offset through the coreboot tables for two
years now, why aren't we using that? (This patch probably does far from
everything that could be done for depthcharge's FMAP handling. But it
does remove another unnecessary point of hardcoded configuration that
needs to be kept in sync.)

BRANCH=None
BUG=None
TEST=Booted Scarlet.

Change-Id: I3b91020993d5461bc4f547e9be0209f9abcf9e87
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/902455
Reviewed-by: Aaron Durbin <adurbin@chromium.org>