vboot/screens: Limit text width to divider width, not to canvas size
We recently reduced the divider width to add padding on the sides of the
canvas, which makes it look better on portrait-sized screens. But some
languages can contain very long strings which are wider than the divider
width. Our existing code can limit these strings to still fit the
screen, but that's not enough anymore. This patch changes it to limit
all strings/images to not exceed the divider width.

BRANCH=None
BUG=None
TEST=Looked at firmware screens on Scarlet.

Change-Id: I0255566e4e51b40f5c710eb7d5720054c43fd165
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/903363
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>