vboot/screens: Add globe symbol to language menu
This patch adds a little globe symbol next to the language menu entry so
that people will be able to find it even if the current language looks
like random squiggles to them. After hours of agitated bikeshedding
between myself and I it was decided that not including this in the menu
option itself for horizontal alignment and selection highlighting
purposes is the better looking option.

CQ-DEPEND=CL:903238
BRANCH=None
BUG=None
TEST=Booted Scarlet, switched language to Malayalam, found my way back.

Change-Id: I02e36314595b0733fc27c161a92fb13b7f542428
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/903365
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>