vboot_aux_fw: Refine error handling
This change makes the error conditions in aux firmware updates
less drastic.  Instead of die() if the firmware is not found
in CBFS return back up the stack so it ignores the update.

Additionally, if the hash check fails when deciding whether to
update just mark that port to be skipped so we do not get stuck
in an endless reboot loop.

BUG=b:74336712
BRANCH=eve
TEST=this was tested manually, first by not including the firmware
update in CBFS, and by changing check_dev_fw_hash() to always fail.

Change-Id: I77d69a76ba9f2a6db97fa2600284806d0ded9c90
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: I22a741cc9a9d016c390d29c0afca54c0a6381a8b
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/985402