Grunt: Switch Stoney SDHCI controller from EMMC to SD card
On updated proto boards and all later boards, the Stoney SDHCI pins will
be connected to the SD card instead of the EMMC part.  This switches it
over in Depthcharge so that we can use the SD card to load recovery
images and boot from it.

BUG=b:72548516
TEST=Boot Grunt to Kernel
BRANCH=None

Change-Id: Iff896bb7168ff19f58052e7c8e75f4618e2ff0ac
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/888095
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Craig Bergstrom <craigb@chromium.org>