vboot: Add init function list
INIT_FUNC handlers (including board_setup) are run before vboot is
initialized, as they may be preparing interfaces for vboot to use.

This means that the board_setup function can't rely on things like
vboot_in_recovery() or access to the GBB flags, as these are not
available until vboot setup has been run.

In order to support this add a list of vboot init functions to be
executed before passing control to vboot.

This provides a path for board_setup to register a function to be
called before vboot is executed to boot a kernel.

BUG=b:67029862
BRANCH=eve
TEST=manual: boot on Eve with subsequent commit that checks both
GBB flags and vboot_in_recovery().

Change-Id: I7bbeba95fed45d6c21456fbf7bf07ca40a3140e3
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: I8bda1de1d1a4fe86a3321ee56f48ec82538ed069
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/985399