nvme: Send shutdown command before reboot/poweroff
Add support for sending a shutdown command to the NVMe device
on a reboot or poweroff cleanup path.

The function will try to send a 'normal' shutdown and wait for
it to complete, but if that times out it will send an 'abrupt'
notification and return immediately.

This is only done on devices that have been enabled, that is if
the reboot happens before depthcharge touches the device it will
not be sent.  With my testing this is ok, I did not see an increase
in the 'unsafe sutdowns' counter if the device had not been touched
by the boot process.

BUG=b:73782738
BRANCH=eve
TEST=check depthcharge initiated reboots before and after storage
has been initialized, in the latter case ensure that the shutdown
command is sent and succeeds and that the 'unsafe shutdowns'
counter does not increase.

Change-Id: I27fa23366aa35ebb8d5807ec02567fcde93b0064
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/959661
Reviewed-by: Furquan Shaikh <furquan@chromium.org>