drivers/bus/spi: Add new bitbanging SPI driver
Same motivation as in coreboot: sometimes you just want to get the
system to boot all the way through while people are still working on the
SPI controller driver. This patchs adds a bitbanging implementation that
can easily be used with a couple of handcrafted functions that implement
depthcharge's GpioOps interface.

BRANCH=None
BUG=None
TEST=Booted Kevin all the way through without SPI controller hardware.

Change-Id: I80132d5cc6498721a2dd956de4290311738189d0
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/982427
Reviewed-by: Aaron Durbin <adurbin@chromium.org>