grunt: Enable support for updating ps8751
Depthcharge now successfully detects the ps8751 and performs firmware
upgrades and downgrades.

The driver code will fail the boot process if the firmware is not found
in cbfs. So we must wait until https://chrome-internal-review.googlesource.com/c/chromeos/overlays/overlay-grunt-private/+/599548
has been merged to merge this patch. Since the change that adds the
firmware to cbfs is only part of the internal repo, any builds using
only the public repos will fail to boot...

BRANCH=none
BUG=b:73498384
CQ-DEPEND=CL:*599548
TEST=Tried downgrading and then upgrading firmware on grunt
Logs for downgrading to 0x34:

    ps8751.1: found SPI flash ID 0x1c11
    ps8751.1: erased 20KB in 440ms
    ps8751.1: verified 0KB in 0s
    ps8751.1: programming 20KB...
    ps8751.1: programmed 20KB in 26s
    ps8751.1: verified 20KB in 18s
    ps8751.1: vendor 0x1da0 product 0x8751 device 0x0001 fw_rev 0x34

    localhost ~ # ectool pdchipinfo 1
    vendor_id: 0x1da0
    product_id: 0x8751
    device_id: 0x1
    fw_version: 0x34

Logs for upgrading to 0x37:

    ps8751.1: found SPI flash ID 0x1c11
    ps8751.1: erased 20KB in 435ms
    ps8751.1: verified 0KB in 0s
    ps8751.1: programming 20KB...
    ps8751.1: programmed 20KB in 26s
    ps8751.1: verified 20KB in 18s
    ps8751.1: vendor 0x1da0 product 0x8751 device 0x0001 fw_rev 0x37

    localhost ~ # ectool pdchipinfo 1
    vendor_id: 0x1da0
    product_id: 0x8751
    device_id: 0x1
    fw_version: 0x37

Change-Id: I18fca21900925f00b635cdc43be8d904aed23370
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/990237
Reviewed-by: Simon Glass <sjg@chromium.org>