grunt: Enable support for updating anx3429
Depthcharge now successfully detects the anx3429 and performs firmware
upgrades

The driver code will fail the boot process if the firmware is not found
in cbfs. So we must wait until https://chrome-internal-review.googlesource.com/c/chromeos/overlays/overlay-grunt-private/+/599770
has been merged to merge this patch. Since the change that adds the
firmware to cbfs is only part of the internal repo, any builds using
only the public repos will fail to boot...

BRANCH=none
BUG=b:73498384
CQ-DEPEND=CL:*599770
TEST=Tried changing the firmware version on grunt
Since there is only one firmware version, I manually modified the .hash file
so depthcharge would think the firmware version was different. This resulted in
depthcharge saying it was flashing the firmware. I can't confirm if it actually
flashed or not since I can't actually change the firmware revision in the bin.

	anx3429.0: active FW header at offset 0x0002 -> 0x0066 0x00d2
	anx3429.0: writing new FW to offset 0x0066, size 0x00d2
	anx3429.0: writing new FW header to offset 0x0002.
	anx3429.0: vendor 0xaaaa product 0x3429 device 0x00ad fw_rev 0x15

	localhost ~ # ectool pdchipinfo 0
	vendor_id: 0xaaaa
	product_id: 0x3429
	device_id: 0xad
	fw_version: 0x15

A normal boot will print the following:

	anx3429.0: vendor 0xaaaa product 0x3429 device 0x00ad fw_rev 0x15

Change-Id: Ifc38d16b5fb0eef1b363024e8a5ca0438927a76c
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/990723
Reviewed-by: Simon Glass <sjg@chromium.org>