[EVE BRANCH ONLY] anx3429: add retry and relax some failures
To reduce user confusion and support calls for ending up in recovery
have failures that are not directly related to updating the ANX FW
non-fatal as they should not leave the FW in a bad state.  Instead
these failures will request an EC reset, which also means they don't
need to try and clean up the PD state on the EC.

Also in this commit I put back a retry for a specific failure that
was seen in testing.

BUG=b:78837043
BRANCH=eve
TEST=manual testing to ensure that failures in these steps result
in an EC reset while update failures go to recovery.

Change-Id: I91db7a5015875f396c785d963a9eebcc1e88d8bf
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1040734
Reviewed-by: Furquan Shaikh <furquan@chromium.org>