drivers/storage/sdhci: Set clock if it is not already enabled
On KBL platforms(poppy/soraka), it is observed that the clock control
register gets reset to 0 after SD card controller is power-gated. In
such cases, it is not right to rely on the previously configured clock
to decide if clock control register needs to be re-programmed. This
change adds a check to see if clock is currently enabled in control
register and if not then it calls sdhci_set_clock even if previously
configured clock value hasn't changed.

BUG=b:111553960
BRANCH=poppy
TEST=Verified by inserting and removing SD card multiple times in
recovery mode and developer mode that it works fine every time.

Change-Id: I1de4885a5ac8fc9c7eaeb590acd8b04fb1c4e502
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1156984
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>