image: Set proper end for FDT buffer, add overlap checks
There are up to three hardcoded memory ranges in depthcharge that are
used at the same time during kernel handoff: the depthcharge code and
data itself, the buffer that vboot loads the kernel into, and (on FDT
systems) the buffer that the FDT is passed to the kernel in. Right now
we just rely on people manually configuring these ranges to not overlap,
and that has been fine so far, but automated checking is always nicer.

This patch ads link-time overlap checks to these regions. On arm64, we
have the additional requirement that all other special memory regions
handled directly by depthcharge should be below the depthcharge code
itself. For simplicity (and since it doubles as a convenient way to
ensure we're not overlapping CBMEM either), just enforce this
requirement across all architectures. The x86 ones (which all use the
same addresses) already conform to this anyway. Some other boards
needed to have their placements slightly adjusted to conform.

Since the FDT area didn't have a well-delimited size for now, let's just
hardcode it to 8MB. Current FDTs are generally in the 50KB range, so
this should be enough forever (famous last words...).

BRANCH=None
BUG=chromium:873135
TEST=None

Change-Id: I52715f13544063549ec07a2adf2f295917298e0e
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1179121
Reviewed-by: Aaron Durbin <adurbin@chromium.org>