image: Expand kernel buffer size to 512MB
This patch expands the kernel buffer to 512MB (or to 128MB on arm32
devices where that's the Linux ABI limit anyway, or to less on some
older specialty platforms where it otherwise wouldn't fit). The kernel
partition size on disk has traditionally only been 16MB, and we don't
plan on changing that... however, on a recovery USB stick we can make
the partition as large as we want, and we may need to do that if we want
to add more stuff to the recovery initramfs. Unfortunately we'll be
stuck with the exisiting limits on old platforms since recovery firmware
is not updateable, but this CL is meant to ensure that we'll at least
have lots of headroom to grow on future platforms.

Some memory layouts had to be adjusted to fit the new buffer. Also
document the individual arch-dependent constraints for placing various
regions in the Kconfig help (some of these only apply to older Linux
versions, but since we don't update our kernels it's better to mention
everything that might apply... boards can still violate them after they
made sure that they're unaffected).

BRANCH=None
BUG=chromium:873135
TEST=Booted Cheza (not feasible to test *every* platform, I've
double-checked everything manually as best as I could).

Change-Id: I23f1a2c78f4e82c1f1e4a16d823ab36d2e90ef1d
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1179122
Reviewed-by: Aaron Durbin <adurbin@chromium.org>