arm64: Read kernel relocation buffer size from kernel image header
If we want to boot potentially very large kernel images, we can no
longer easily hardcode the size of the kernel relocation buffer on
arm64. Instead, just take the image_size provided for this purpose in
the kernel header and find a buffer that fits exactly that size.

BRANCH=None
BUG=chromium:873135
TEST=Booted Cheza.

Change-Id: Ia03a206eb96a92cf4b076d68cdc1e797510436d9
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1179267
Reviewed-by: Aaron Durbin <adurbin@chromium.org>