vboot-aux-fx: don't fail if reboot required
If the TCPC firmware upgrade reports that it needs a reboot first, do
not consider that a failure mode. This allow the firmware upgrade
process to continue on a subsequent boot.

This happens with the EC firmware is already in RW and the i2c tunnel
protect is already engaged. The only existing way to un-protect the
tunnel is to reboot.

BRANCH=none
BUG=b:112722728
TEST=Jump EC to RW, then apshutdown, then powerb. Verify that firmware
upgrade portion does not cause recovery, instead it causes a reboot and
the subsequent boot will work correctly (either upgrade or skip)

Change-Id: I96342cd47de103d4a33595bd1c3331709d781781
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1200316
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
(cherry picked from commit 861cc3c03cfa5017d108fefddcb6672a036c4daa)
Reviewed-on: https://chromium-review.googlesource.com/1229181
Reviewed-by: Martin Roth <martinroth@chromium.org>
Commit-Queue: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>