device_tree: Change property name parameters to const char *
Some dt_add_<dtype>_prop() functions use char * as datatype for the
property name parameter. They value is assigned to the 'name' field of
struct FdtProperty, which is of type const char *. Change the parameter
type to const char * to allow constant strings to be passed to
dt_add_<dtype>_prop().

BUG=b:113114217
TEST=see CL:Ib49c5e5577a0ca2a52fbb2dde2a25580397464ff
BRANCH=none

Change-Id: Ica32f6449cf10d1e3991ed7f09d8551578c43e1f
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1249702
Reviewed-by: Julius Werner <jwerner@chromium.org>