device_tree: Have absolute paths start with '/'
Currently DT paths are *not* expected to start with '/'. This is not
what the spec says (see Devicetree Specification v0.2, 2.2.3 Path Names)
and also not what is done by Linux.

Change dt_find_node_by_path() to expect paths to start with '/' and add
a leading '/' to all DT path strings. Besides the compatibility with the
spec this change is also needed to support aliases in the future.

BUG=b:113114217
TEST=build for cheza and boot
BRANCH=none

Change-Id: If54bc1428b891525f39690ea60c81a5b6102990a
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1252770
Reviewed-by: Julius Werner <jwerner@chromium.org>
(cherry picked from commit e2e51c6b33b12b34ed554513daec04519213bf64)
Reviewed-on: https://chromium-review.googlesource.com/c/1277528
Reviewed-by: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>