cheza: Add DT patch entries for legacy VPD names for MACs
The ODM screwed up and programmed 'bluetooth_mac' and 'wifi_mac' in
the VPD. Add entries for these too.

This will result in warnings when either entries are not found. The
patch can be reverted when all/most systems with the legacy entries
have been deprecated or migrated to the correct names

BUG=b:113114217
TEST=see CL:Ib49c5e5577a0ca2a52fbb2dde2a25580397464ff
BRANCH=none

Change-Id: I87a9cd46be6dd6bdfe4e7e6091341fd6b81ad14e
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1252772
Reviewed-by: Julius Werner <jwerner@chromium.org>