Implement the menu screen for alternative firmware
Show a new screen with a list of bootloaders which can be selected using
the number keys. The list is displayed by depthcharge, in English only.

Item 0 is the default but always exists as a separation option, so don't
show it in the menu.

Also adjust the list adding algorithm, so that items remain in the order
set by chromeos-bootimage.

CQ-DEPEND=CL:1273268
BUG=chromium:837018
BRANCH=grunt
TEST=emerge-grunt --nodeps depthcharge
With other changes, show the menu with Ctrl-L, exit with Esc and select
a bootloader with keys 0, 1 or 2

Change-Id: I2ce95c6d6a873934978074b7bca665c82e31871d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1273269
Reviewed-by: Julius Werner <jwerner@chromium.org>