anx3429: auto-correct OTP header bit flips
this code detects when there are a small number of bits "unset" that
should be set in the OTP header section.  in this section, an entry is
void if all bits are set to 1.  sometimes, not all bits are set and
these situations are easy to detect so we try to correct them by
re-writing the header.

also, if an entry seems valid but is implausible, we void it out.
this also has the side effect that a fresh entry will be used and
written correctly.

BUG=b:118690201
BRANCH=none
TEST=verified anx3429 firmware update on reef

Change-Id: I79193382e7a50bbde8d354ce72e22d1402d03e0b
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1314295
Commit-Ready: Caveh Jalali <caveh@google.com>
Tested-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>