anx3429: auto-correct OTP header bit flips
this code detects when there are a small number of bits "unset" that
should be set in the OTP header section.  in this section, an entry is
void if all bits are set to 1.  sometimes, not all bits are set and
these situations are easy to detect so we try to correct them by
re-writing the header.

also, if an entry seems valid but is implausible, we void it out.
this also has the side effect that a fresh entry will be used and
written correctly.

BUG=b:118690201
BRANCH=none
TEST=verified anx3429 firmware update on reef

Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 3357cc00dabf3beab1eb6150db56f60a0a324c24
Original-Change-Id: I79193382e7a50bbde8d354ce72e22d1402d03e0b
Original-Signed-off-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/1314295
Original-Commit-Ready: Caveh Jalali <caveh@google.com>
Original-Tested-by: caveh jalali <caveh@chromium.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
(cherry picked from commit 3357cc00dabf3beab1eb6150db56f60a0a324c24)

Change-Id: Icccff010d49ed2e236fd4f1cbdd5bf4599c6455a
Reviewed-on: https://chromium-review.googlesource.com/c/1338364
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Duncan Laurie <dlaurie@google.com>
Tested-by: Duncan Laurie <dlaurie@google.com>
Trybot-Ready: Duncan Laurie <dlaurie@google.com>