flags: Ignore FLAG_PWRSW for detachable UI
In the detachable UI, we always treat the power button as a keyboard
key. This means that the existing VbExIsShutdownRequested() path to
check the power button is not needed, but it still runs and depthcharge
still enforces that flag to be installed. Due to this all boards have to
install a non-functional mock power button flag somewhere.

This patch skips the power button check for detachables so that we can
stop working around this framework quirk.

BRANCH=None
BUG=None
TEST=Booted Cheza (with pending patches to make that actually work).

Change-Id: Ieacf64cd140eaa61322a5c654a65ca3ace3f99ff
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1363875
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: T Michael Turney <mturney@codeaurora.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>