tpm: fix a casting bug in vb2ex_tpm_get_mode
enum vb2_tpm_mode* is casted to uint8_t* and then written to by
tpm_get_mode.  tpm_get_mode only writes to the least significant
byte of the enum value, and the rest is left untouched.  Depending
on the endianness of the architecture in question, the byte written
to in the enum may also be incorrect.

As an example, the TPM mode may send back 0x00 for mode_val (which
corresponds to ENABLED_TENTATIVE).  But the enum vb2_tpm_mode
value could end up as: 0x02 0x73 0xca 0x00.

BUG=b:120587305
TEST=recompile image, flash to eve, and ensure TPM mode can be
     retrieved properly on reboot.

Change-Id: I7bd398543a599ae0470e98121bd4f4ffc4b8ad68
Reviewed-on: https://chromium-review.googlesource.com/c/1367344
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>