payload: Adjust the API so that cbfs_media is internal
At present the payload functions take a cbfs_media * as an argument. But
at present we use the same area for all payload operations, so this is
unnecessary.

Add a new function to return the cbfs_media struct, and keep a single one
of these in payload.c to simplify the API.

BUG=b:120555612
BRANCH=grunt
TEST=manually test on grunt, pressing '1' to boot into U-Boot

Change-Id: I1f7eaaac185c5233760cee5e8b1d8cc6c9080524
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1370604
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Signed-off-by: Martin Roth <martinroth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1396162