drivers/flash: Add fast_spi driver for Intel HWSEQ
In order to support reading from SPI locations that are not memory
mapped on an Intel SOC we need to use the chipset HWSEQ interface.

This driver is ported directly from coreboot with minimal changes
to fit into the coding style for depthcharge.

Additionally the flash regions of the SPI are read out of the
descriptor and put into a table for depthcharge to access.  This
will be used by subsequent drivers.

BUG=b:120628933
BRANCH=none
TEST=successfully replace mmio flash interface on sarien

Change-Id: I077a9a5193170c35e678e2975a70d77ee3759aae
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1378389
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>