cr50_i2c: Add logging for cr50_i2c_tpm_ready()
This function may fail. Add logging to detect this condition even though
we cannot report it to the caller.

The cr50_i2c_write() call inside can also fail. For that case we have
already logged the error, and adding a short delay after the failure
should not make anything worse.

BUG=b:113880780
BRANCH=grunt
TEST=FW_NAME=careena USE="em100-mode" emerge-grunt --nodeps depthcharge

Change-Id: I810243dd92bd2eabf1fb60c13d7bf56aeba1cb14
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1384574
Reviewed-by: Raul E Rangel <rrangel@chromium.org>