multiboot: Fix multiboot mod_end value when bootdata is used
When bootdata_prepare() adds contents into the ZBI (nee bootdata) format
ramdisk, it adjusts the size of the ramdisk.  But this was not being
represented in the Multiboot info describing the bounds of the ramdisk.
So Zircon's Multiboot loader could not impose sanity checks that the size
of the image described by the ZBI headers inside the image itself fit in
the size that Multiboot says was actually loaded.

BUG=https://crbug.com/917455
BRANCH=eve
TEST=manual testing with bootdata format kernel

Change-Id: Ib3a23fcc9c135506c50513c6dbf6e88425a79663
Reviewed-on: https://chromium-review.googlesource.com/1387968
Commit-Ready: Duncan Laurie <dlaurie@google.com>
Tested-by: Duncan Laurie <dlaurie@google.com>
Reviewed-by: Duncan Laurie <dlaurie@google.com>