COMMIT-QUEUE: Replace Gru Pre-CQ target with Scarlet
Some change in biod seems to have broken Gru builds. A fix is on the
way, but since Gru isn't a "real" board we should probably stop relying
on it anyway. Add Scarlet in its place so that we're still testing at
least one target with SPI-based Cr50.

BRANCH=None
BUG=chromium:921654
TEST=None

Change-Id: Ibe5de9b3d7f5b599d7cefccfb7b54261fa3089fd
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1410068
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>