sarien: Enable recovery mode GPIO to be resampled
By default the recovery GPIO is not re-sampled at runtime.

In order to support FAFT we need to be able to read the recovery
button to confirm the switch to developer mode.

Eventually this will also be able to read the power button press
from H1 but this will allow FAFT tests to work.

BUG=b:121380403
BRANCH=none
TEST=test_that DUT firmware_DevMode

Change-Id: Ic72dd0076fbd5c77a9084f603ddf617d61f3785e
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1428302
Reviewed-by: Furquan Shaikh <furquan@chromium.org>