ec/wilco: Use alternate reboot method if needed
By default the EC reboot code uses a command that might not
be available depending on the current EC trust level.  If it
returns an "access denied" error instead configure the EC to
reboot after the host powers off, which is the next step that
vboot will do in this path.

This fixes an issue where the system could remain powered off
if this command was not available.

BUG=b:120628933
BRANCH=none
TEST=ensure system always reboots after RO reboot request

Change-Id: Ic9f45b21b3c70b0d20411e75fde4d0f5295136e8
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1458479
Reviewed-by: Furquan Shaikh <furquan@chromium.org>