ec/wilco: Add support for EC software sync
This change adds the flash update interface for the Wilco EC and
the necessary hooks for it to be updated from vboot.

Currently the EC hash is done by directly hashing the image in SPI.

BUG=b:120628933
BRANCH=none
TEST=manual testing of EC software sync on Arcada to ensure that
it can detect and update the EC image from what is found in cbfs.

Change-Id: I143e99cfc27084c186d6fd4b141f8db41b952beb
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1458481
Reviewed-by: Furquan Shaikh <furquan@chromium.org>