depthcharge/legacy: add SHA-256 hash verification for RW_LEGACY
Adds option to verify a payload in legacy before running it.
This is mostly a cherry-pick of CL:1127904.

The scheme involves looking up a hash file that has ".sha256" appended
to the name of the payload.  Alternatively the hash name could be added
to the "altfw/list" file that contains the list of firmwares but we'd
have to deal with the prior file format (and it could make it easier for
someone to swap out one hashed payload for another and run that other
payload instead).

BUG=b:124358784
BRANCH=None
TEST=Used with a subsequent change to confirm that rom can be
successfully verified, and that when an intentionally invalid hash is
used the mismatch is detected and payload isn't run.
CQ-DEPEND=CL:1471051, CL:1470548

Change-Id: Ib70e23fc8073066412628321c79e5d15adcaac33
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1471053
Reviewed-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>