vboot_ec: always suppress power button on x86
vboot_select_and_load_kernel() disables the power button on detachables,
and update_vboot_aux_fw() has analogous functionality to only disable and
restore the power button on non-detachables.

I've got an upcoming use case that wants to temporarily disable the power
button so the status of the button can be queried.  This would further
complicate the conditions on when to enable or disable the power button.

The screen/menu handling in depthcharge already has functionality to
monitor the state of the power button for detachables  This
functionality is also suitable for non-detachables, so this change opts
to disable the power button for all x86 boards.

BUG=b:124358784
BRANCH=none
TEST=build and verified that method can be used to disable and re-enable
the power button.  Verified that power button can be used at recovery
screen and dev screen to turn off Atlas and Arcada.

Change-Id: I43fd80fe9adeb797050808ab3b0001fb69abc35f
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1470272
Reviewed-by: Julius Werner <jwerner@chromium.org>