vboot: add power button query functionality
The primary purpose of this change is to add a flag to VbExGetSwitches()
so the current user physical presense (via the status of the power button)
can be queried.

The flags have also been renamed from the historical _INIT_ use to
_SWITCH_ to reflect the current/actual API they're used with.

BUG=b:124358784
BRANCH=none
TEST=Locally built and flashed. Confirmed that the power button status
can be queried when the power button is disabled.
CQ-DEPEND=CL:1471190

Change-Id: I652a3ee3d62fe77ad8de6f64c7844f63cf536e47
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1470273
Reviewed-by: Julius Werner <jwerner@chromium.org>