drivers/ec/wilco: Put keyboard into translated mode
This keyboard controller does not come up in translated mode by
default, so if it is not enabled in firmware then it will never
work properly in the Linux kernel because it will be instantiated
as an "AT Raw" keyboard instead of an "AT Translated" keyboard.

In order to be able to remove the keyboard initialization in
firmware we need to set this bit in the command register before
booting to the OS.

It is necessary to call i8042_probe() first or the later functions
will not succeed.

BUG=b:126633269
BRANCH=none
TEST=boot on sarien and ensure keyboard works in the OS.

Change-Id: I394ab1063ce00ef1df3b42c93ff2ba5c56761eec
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1491882
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>