vboot/screens: Get rid of video console, use HWID font for everything
This patch changes VbExDisplayDebugInfo() to use the HWID font (which
now includes all printable characters) and overall make it a little
prettier. This allows us to completely eliminate the dependency on the
libpayload video console and the corebootfb framebuffer driver for
normal images. corebootfb and cbgfx have always been sort of parallel
implementations for roughly the same stuff that both write to the
framebuffer directly, so it's cleaner to only use one of them for
everything.

Fastboot, netboot and CLI images continue to use the video console for
specific developer use cases (which is probably fine), but production
images no longer pull any of it in (which saves about 3KB/3% of the
compressed payload binary on my Kevin).

BRANCH=None
BUG=None
TEST=Played around with TAB output on Kevin, both with and without
bitmaps in CBFS, configured for both normal and detachable UI.

Change-Id: Iea709f97ae53a279f3e48decc56de84da4568845
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1496905
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Mathew King <mathewk@chromium.org>