depthcharge: Write vendor data to ro vpd
This change does two things, first the SRWD bit must be 0 and the vendor
data contents must be all spaces in order for the vendor data to be
settable in the vboot ui. Next when VbExSetVendorData is called the
current VPD is read from flash, modified with the new data and written
back to flash.

This change does not enable the SRWD bit because whenever the system
boots with SRWD cleared all of the block protect bits are also cleared.
I'm not sure why the block protect bits are being cleared and if they
are not set correctly the RO firmware will not be protected.

BUG=b:124297157
TEST=Verified on an Arcada device that if serial_number is set to all
     spaces and SRWD is not set then a new serial number can be entered
     and read successfully from userspace using "vpd -l".  Also verified
     that the serial number cannot be set if it is already set or SRWD
     is set.
BRANCH=none

Change-Id: I1b8d19d19e7bc6d93dc369ab53ee658782838077
Signed-off-by: Mathew King <mathewk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1483996
Reviewed-by: Julius Werner <jwerner@chromium.org>