Add flash operation to set wp
The flash_set_wp_enabled function will use the protection_mapping which
must be defined in board.c in order to use. FlashProtectionMapping
defines the status register value that will set both the block
protection bits and the write protection bit for a flash with a given
id.

This also adds a function to the FlashOps to read the id of the flash.

BUG=b:124297157
TEST=Verified on Arcada with other patches in stack
BRANCH=none

Change-Id: I0db28fe92373494a727726b989ac8705db01b63a
Signed-off-by: Mathew King <mathewk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1516678
Reviewed-by: Julius Werner <jwerner@chromium.org>