devicetree: Match debug output format to dtc -O dts output
This patch updates the device tree dumping functions (not compiled by
default but available for debugging) to output properties and nodes in a
format similar to .dts files that is very close to what dtc outputs when
you decompile a .dtb with it. This makes it easier to match device tree
dumps in depthcharge with device tree dumps generated by other device
tree tooling.

BRANCH=None
BUG=None
TEST=Dumped a couple of overlaid device trees that included both strings
and binary data.

Change-Id: I1a450141d151b25828011aac3fed9527b42d59c2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1536386
Reviewed-by: Hung-Te Lin <hungte@chromium.org>