drivers/ec/wilco: Add support for PD firmware update
The PD firmware update interface is very similar to the EC update
interface, but different enough that it needs a separate driver.

TCPC types/vendors can have a different sub-type in the flash inventory
and use a separate set of EC sub-commands.

The PD update check is based on the version read from the EC flash
inventory command.

The EC command timeout is increased to handle the case of an interrupted
flash taking a very long time to respond to the 'flash start' command
in order to recover and reflash successfully.

BUG=b:123224590
BRANCH=none
TEST=using two differenet PD FW versions test updating multiple times
and ensure that the process is successful.

Change-Id: I46a5f3258920dd0f0d375c5296859638396164b1
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1538901
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>